+++ CHANGES.rst
@@ -5,6 +5,10 @@ grafana-wtf changelog
 
 in progress
 ===========
+Improve the situation with large Grafana installations, see #2.
+Thanks `@jangaraj <https://github.com/jangaraj>`_.
+
+- Add option ``--cache-ttl`` for controlling the cache expiration time
 
 
 2019-04-21 0.3.1