+++ CHANGES.rst
@@ -5,11 +5,16 @@ grafana-wtf changelog
 
 in progress
 ===========
-Improve the situation with large Grafana installations, see #2.
+
+
+2019-05-07 0.4.0
+================
+Slightly improve the situation with large Grafana installations, see #2.
 Thanks `@jangaraj <https://github.com/jangaraj>`_.
 
 - Add option ``--cache-ttl`` for controlling the cache expiration time
 - Improve error logging when hitting Grafana unauthorized
+- Improve performance of search routine
 
 
 2019-04-21 0.3.1